The Incentive-Deterrence Model

Basics

The dynamics of the system is,

\[\left\{\begin{array}{ll} y_{t+1} & = \varphi_{t+1}(x_t, y_t, \tau), \\ x_{t+1} & = \Gamma x_k - \hat\Gamma y_{t+1} + \lambda_t. \end{array}\right.\]

The main fixed-point iteration is at,

Criminos.F! — Function

F!(
    vector_ms::Vector{MarkovState{R,Tx}},
    vec_Ψ::Vector{BidiagSys{Tx,Tm}};
    fₜ=Criminos.decision_identity, targs=nothing, # function of decision
    fₘ=Criminos.mixed_in_identity, margs=nothing, # function of mixed-in effect
    kwargs...
) where {R,Tx,Tm}

The F! function performs a fixed-point iteration on a vector of MarkovState objects and a vector of BidiagSys objects. It updates the states in vector_ms and vec_Ψ according to the specified functions fₜ, fₘ, and Fₓ.

Arguments

vector_ms::Vector{MarkovState{R,Tx}}: A vector of MarkovState objects.

vec_Ψ::Vector{BidiagSys{Tx,Tm}}: A vector of BidiagSys objects.

fₜ: A function that updates the states in vector_ms collectively. Default is Criminos.decision_identity.

targs: Additional arguments for the function fₜ. Default is nothing.

fₘ: A function that updates the states in vec_Ψ collectively. Default is Criminos.mixed_in_identity.

margs: Additional arguments for the function fₘ. Default is nothing.

kwargs...: Additional keyword arguments.

We define two data structs:

the system struct, $\Psi$, to store constants

the state variable, $z$, for $(x,y)$ in the math, and attributes to enable efficient computations 😍

Criminos.BidiagSys — Type

mutable struct BidiagSys{Tx,Tm}

Bidiagonal system in the paper.

Fields

n::Int: Number of states.

γ::Tx: Rate retention.

λ::Tx: Rate arrival.

q::Tx: Probability.

M::Tm: Helper matrix.

Q::Tm: Helper matrix.

Γ::Tm: Helper matrix.

Γₕ::Tm: Helper matrix.

style::Symbol: Random or known.

Constructors

BidiagSys(n::Int; style=:rand): Constructs a BidiagSys object with the specified number of states n and style style.

Criminos.MarkovState — Type

struct MarkovState{R,Tx}

A mutable struct representing the state of a Markov-type Dynamical system

Fields

k::Int: The iteration number.

n::Int: The number of states.

z::Tx: The fix-point iterate.

x::Tx: The iterate.

x₋::Tx: The previous iterate.

ρ::Tx: The probability.

y::Tx: The recidivists.

y₋::Tx: The previous recidivists.

τ::Tx: The treatment probability.

f::Real: The objective value of the mixed-in function.

Constructors

MarkovState(k, n::Int; z=[Random.rand(Float64, n); Random.rand(Float64, n)], τ=Random.rand(Float64, n), β::Real=1.0): Constructs a MarkovState object with the given parameters.

Dynamics of $x$

The $x$ part is simple,

Criminos.Fₓ — Function

population dynamics x

\[x_{t+1} = \Gamma x_t - \hat\Gamma y_t + \lambda_t.\]

In this view, we use $\Psi$

\[\quad \Gamma =\mathrm{diag}(\gamma), \quad \hat \Gamma = \begin{bmatrix} \gamma_0 & 0 & \cdots & 0 \\ -\gamma_0 & \gamma_1 & \cdots & 0 \\ \vdots & \ddots & \ddots & \vdots \\ 0 & \cdots & -\gamma_{N-2} & \gamma_{N-1}, \end{bmatrix}.\]

such that we call the bi-diagonal system. The associated "potential" or Lynapunov function is,

\[u(x, y) = \frac{1}{2}(x)^T(I - \Gamma)x - (\lambda)^T x + (x)^T \hat{\Gamma} y\]

Adversarial Energy, and Best Response

Note $\varphi$ of $y$ follows the utility/potential/energy $\omega$.

Criminos.w — Function

linear quadratic function

Criminos.mixed_in_gnep_best! — Function

best response for the GNEP problem 
```math
y \in [0, x_{k+1}]
```

Criminos.mixed_in_gnep_grad! — Function

"better" response for the GNEP problem 
```math
y \in [0, x_{k+1}]
```

we take arguments

∇₀, H₀, ∇ₜ, Hₜ, _... = args

for any $y,\tau$, compute $c$ from the exogenous arrivals

\[c = Q\lambda\]

the deterrence & incentives

\[\begin{aligned} \nabla^2\omega &= \operatorname{diag}(\tau) \cdot H_t \cdot \operatorname{diag}(\tau) + H_0 \\ \nabla\omega &= (\nabla^2\omega) y + \underbrace{\nabla_t \tau - \nabla_0 c}_{g} + \hat \Gamma^Tx \end{aligned}\]

and finally

\[\omega(y,\tau) = \frac{1}{2}y^T (\nabla^2\omega) y + y^T \left(\nabla_t \tau - \nabla_0 c\right) + y^T\hat \Gamma^Tx\]

Note

One can see $\tau$ is crucial to compute the second-order derivatives. If $\tau = 0$, this reduce to $H_0, \nabla_0 c$... This can be taken as intrinsic property of a subpopulation

Extension to Multiple Subpopulation

For multiple population, $\nu = 0,..., M-1$, now $\omega$ has multiple blocks, $y^0, \ldots, y^\nu, \ldots, y^{M-1}$, stick to the parameterization,

\[g^\nu = \nabla^\nu_t \tau^\nu - \nabla^\nu_0 c^\nu\]

Then the energy can we rewritten as,

\[\begin{aligned} \omega(y, x, \tau) := y^T \begin{bmatrix} (\hat{\Gamma}^0)^T x^0 \\ \vdots \\ (\hat{\Gamma}^{M-1})^T x^{M-1} \\ \end{bmatrix} + y^T \begin{bmatrix} g^0 \\ \vdots \\ g^{M-1} \\ \end{bmatrix} + \frac{1}{2} y^T \begin{bmatrix} H^1 & \cdots & H^{1,M-1} \\ \vdots & \ddots & \vdots \\ H^{M-1,1} & \cdots & H^{M-1} \\ \end{bmatrix} y \end{aligned}\]

Note

If $H_t^{i,j} \neq 0$, then the recidivism of $i$ is affected by the recidivism of $j$.

Population is still independent,

\[\forall \nu, \quad u^\nu(x^\nu, y_k^\nu) = \frac{1}{2}(x^\nu)^T(I - \Gamma^\nu)x^\nu - (\lambda^\nu)^T x^\nu + (x^\nu)^T \hat{\Gamma}^\nu y_k^\nu\]